completed Space Shotter game
Characters : Tom And Jerry
Levels        : Unlimited
Powers      :  Triple Bullets

Total there are 3 screen one is for main screen to select our character and second is playing window and last is screen is to display high score and our score and to go back to home screen or space to continue playing.

Thankyou,

Done by Ajay Sharon.